Fix move validation
Piece getMove method now takes in the ChessGameState as a parameter instead of
the board, allowing direct access to the chessboard and its associated helper methods.

ChessHumanPlayer now validates the requested move inside the onTouch method. If valid, the
move is sent over to the game and processed as normal. If invalid, the selected piece is
invalidated and nothing is changed.