Add check detection
Added new method getChecks to each Piece.java subclass to find all moves that will result in a
check of the King. Currently not made for the King itself as more functionality is needed.

Currently can determine if the king is in check only after the controller selects the king piece. Otherwise,
it will not force the king to move out of a checked position. Functionality in production.